Split widget: mini chart
